move oam serialization into its own class
